Refactor ods
